\name{Rgeogram}

\alias{Rgeogram}
\alias{Rgeogram-package}

\docType{package}

\title{An R interface to the Geogram library}

\description{
The package provides an interface to the Geogram C++ library, with methods to solve semi-discrete optimal transport problems in two or three dimensions.
}

\details{
\tabular{ll}{
Package: \tab Rgeogram\cr
Type: \tab Package\cr
Version: \tab 0.3.0\cr
Date: \tab 2018-09-15\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab Yes\cr
Depends: \tab Rcpp, RcppArmadillo \cr }
}

\section{How to cite this package}{
 \preformatted{@Manual{LBGOS2018,
 author       = {Bruno Levy, Odran Bonnet, Alfred Galichon, Keith O'Hara, Matt Shum},
 title        = {{Rgeogram}: An R interface to the Geogram library},
 year         = {2018},
 note         = {R package version 0.3.0.}}}
\emph{}
}

\section{License}{
Rgeogram is licensed under the GNU General Public License (GPL) version 2, or (at your option) any later version.
}

\author{Keith O'Hara}