\name{Rgeogram}

\alias{Rgeogram}
\alias{Rgeogram-package}

\docType{package}

\title{R interface to Geogram}

\description{
The Rgeogram package is an R interface to the Optimal Transport methods in Geogram.
}

\details{
\tabular{ll}{
Package: \tab Rgeogram\cr
Type: \tab Package\cr
Version: \tab 0.2.0\cr
Date: \tab 2018-08-05\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab Yes\cr
Depends: \tab Rcpp, RcppArmadillo \cr }
}

\section{How to cite this package}{
 \preformatted{@Manual{OHara2018,
 author       = {Keith O'Hara},
 title        = {{Rgeogram}: R interface to Geogram.},
 year         = {2018},
 note         = {R package version 0.2.0.}}}
\emph{}
}

\section{License}{
Rgeogram is licensed under the GNU General Public License (GPL) version 2, or (at your option) any later version.
}

\author{Keith O'Hara}